+++ setup-env.sh
@@ -2,7 +2,8 @@
 
 echo "$(date): Starting setup-env.sh"
 
-export CDK_VERSION=1.83.0
+# this is also set in cdk/setup.py
+export CDK_VERSION=1.87.1
 export PACKER_VERSION=1.5.5
 export TASKCAT_VERSION=0.9.21